+++ src/store.ts
@@ -44,7 +44,7 @@ import {
   CustomResolverMap,
 } from './data/readFromStore';
 
-import assign = require('lodash.assign');
+import assign = require('lodash/assign');
 
 export interface Store {
   data: NormalizedCache;
@@ -110,7 +110,7 @@ export function createApolloReducer(config: ApolloReducerConfig): Function {
 
       return newState;
     } catch (reducerError) {
-      return Object.assign({}, state, {
+      return assign({}, state, {
          reducerError,
        });
     }
@@ -148,6 +148,9 @@ export function createApolloStore({
   // XXX to avoid type fail
   const compose: (...args: any[]) => () => any = reduxCompose;
 
+  // Note: The below checks are what make it OK for QueryManager to start from 0 when generating
+  // new query IDs. If we let people rehydrate query state for some reason, we would need to make
+  // sure newly generated IDs don't overlap with old queries.
   if ( initialState && initialState[reduxRootKey] && initialState[reduxRootKey]['queries']) {
     throw new Error('Apollo initial state may not contain queries, only data');
   }